Microservice2/requirements.txt
fastapi
uvicorn
faiss-cpu  # or faiss-gpu if you're using GPU
numpy
requests
openai  # If you're using OpenAI's API
pydantic

